+++ actionpack/lib/action_controller/metal.rb
@@ -135,7 +135,7 @@ module ActionController
     end
 
     def self.make_response!(request)
-      ActionDispatch::Response.new.tap do |res|
+      ActionDispatch::Response.create.tap do |res|
         res.request = request
       end
     end